Maze Escape
Maze Escape is a popular 1 player game where a bot is trapped in a maze and is expected to find its way out. In this version of the game, 2 bots whose positions are not visible to each other are placed at random points in the maze. the first bot to find its way out of the maze wins the game.

https://www.hackerrank.com/challenges/maze-escape/problem

This problem was solved by me using an algorithm for the robot to always keep the wall to its 'right' until it escapes.